Move Galley test data under galley/testing, for sharing among multiple different tests.
