+++ galley/pkg/testing/testdata/dataset/config.istio.io/v1alpha2/circonus.yaml
