+++ galley/pkg/testing/testdata/dataset/extensions/v1beta1/ingress_basic_meshconfig.yaml
