+++ galley/pkg/testing/testdata/dataset/networking.istio.io/v1alpha3/gateway.yaml
